# web/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateResearchTaskRequest": {
        "properties": {
          "client_request_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Client Request Id",
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/ResearchTaskOptions"
          },
          "query": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "client_request_id",
          "query"
        ],
        "title": "CreateResearchTaskRequest",
        "type": "object"
      },
      "CreateTaskResponse": {
        "properties": {
          "links": {
            "$ref": "#/components/schemas/TaskLinks"
          },
          "task": {
            "$ref": "#/components/schemas/TaskSnapshot"
          }
        },
        "required": [
          "task",
          "links"
        ],
        "title": "CreateTaskResponse",
        "type": "object"
      },
      "DeepSeekSettingsResponse": {
        "properties": {
          "api_key_configured": {
            "title": "Api Key Configured",
            "type": "boolean"
          },
          "api_key_required": {
            "title": "Api Key Required",
            "type": "boolean"
          },
          "default_model": {
            "title": "Default Model",
            "type": "string"
          },
          "provider": {
            "const": "deepseek",
            "default": "deepseek",
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "default_model",
          "api_key_required",
          "api_key_configured"
        ],
        "title": "DeepSeekSettingsResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResearchEvent": {
        "properties": {
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "level": {
            "$ref": "#/components/schemas/EventLevel",
            "default": "info"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "$ref": "#/components/schemas/ReadProgressPayload"
              }
            ],
            "title": "Payload"
          },
          "schema_version": {
            "default": 1,
            "minimum": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "sequence": {
            "minimum": 1,
            "title": "Sequence",
            "type": "integer"
          },
          "stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchStage"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "task_id": {
            "minLength": 1,
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "sequence",
          "event_type"
        ],
        "title": "ResearchEvent",
        "type": "object"
      },
...
      "ResearchStage": {
        "enum": [
          "orchestrate",
          "search",
          "filter",
          "read",
          "analyze",
          "synthesize",
          "critic"
        ],
        "title": "ResearchStage",
        "type": "string"
      },
      "ResearchTaskOptions": {
        "properties": {
          "max_papers": {
            "default": 15,
            "maximum": 15.0,
            "minimum": 3.0,
            "title": "Max Papers",
            "type": "integer"
          },
          "output_language": {
            "default": "auto",
            "enum": [
              "auto",
              "zh-CN",
              "en"
            ],
            "title": "Output Language",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/AcademicSource"
            },
            "minItems": 1,
            "title": "Sources",
            "type": "array"
          }
        },
        "title": "ResearchTaskOptions",
        "type": "object"
      },
      "ResearchTaskResultResponse": {
        "properties": {
          "analysis": {
            "additionalProperties": true,
            "title": "Analysis",
            "type": "object"
          },
          "analysis_findings": {
            "items": {
              "$ref": "#/components/schemas/AnalysisFindingResponse"
            },
            "title": "Analysis Findings",
            "type": "array"
          },
          "capabilities": {
            "$ref": "#/components/schemas/ResultCapabilities"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationResponse"
            },
            "title": "Citations",
            "type": "array"
          },
          "critique": {
            "additionalProperties": true,
            "title": "Critique",
            "type": "object"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceResponse"
            },
            "title": "Evidence",
            "type": "array"
          },
          "paper_claims": {
            "items": {
              "$ref": "#/components/schemas/PaperClaimResponse"
            },
            "title": "Paper Claims",
            "type": "array"
          },
          "paper_insights": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Paper Insights",
            "type": "array"
          },
          "papers": {
            "items": {
              "$ref": "#/components/schemas/ResearchPaperResult"
            },
            "title": "Papers",
            "type": "array"
          },
          "partial": {
            "default": false,
            "title": "Partial",
            "type": "boolean"
          },
          "report_markdown": {
            "default": "",
            "title": "Report Markdown",
            "type": "string"
          },
          "research_plan": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Research Plan",
            "type": "array"
          },
          "run_metadata": {
            "additionalProperties": true,
            "title": "Run Metadata",
            "type": "object"
          },
          "statistics": {
            "additionalProperties": true,
            "title": "Statistics",
            "type": "object"
          },
          "structured_report": {
            "$ref": "#/components/schemas/StructuredReportResponse"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "task_id"
        ],
        "title": "ResearchTaskResultResponse",
        "type": "object"
      },
...
      "StageSnapshot": {
        "properties": {
          "attempt": {
            "default": 1,
            "minimum": 1.0,
            "title": "Attempt",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "duration_ms": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "stage": {
            "$ref": "#/components/schemas/ResearchStage"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/StageStatus",
            "default": "pending"
          }
        },
        "required": [
          "stage"
        ],
        "title": "StageSnapshot",
        "type": "object"
      },
...
      "TaskLinks": {
        "properties": {
          "events": {
            "title": "Events",
            "type": "string"
          },
          "result": {
            "title": "Result",
            "type": "string"
          },
          "self": {
            "title": "Self",
            "type": "string"
          }
        },
        "required": [
          "self",
          "events",
          "result"
        ],
        "title": "TaskLinks",
        "type": "object"
      },
      "TaskProgress": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "metrics": {
            "additionalProperties": true,
            "title": "Metrics",
            "type": "object"
          }
        },
        "title": "TaskProgress",
        "type": "object"
      },
      "TaskSnapshot": {
        "properties": {
          "available_actions": {
            "items": {
              "type": "string"
            },
            "title": "Available Actions",
            "type": "array"
          },
          "client_request_id": {
            "title": "Client Request Id",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "effective_locale": {
            "default": "en",
            "enum": [
              "zh-CN",
              "en"
            ],
            "title": "Effective Locale",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_sequence": {
            "default": 0,
            "minimum": 0.0,
            "title": "Last Sequence",
            "type": "integer"
          },
          "options": {
            "additionalProperties": true,
            "title": "Options",
            "type": "object"
          },
          "parent_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Task Id"
          },
          "progress": {
            "$ref": "#/components/schemas/TaskProgress"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "replay_events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Replay Events",
            "type": "array"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageSnapshot"
            },
            "title": "Stages",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "statistics": {
            "additionalProperties": true,
            "title": "Statistics",
            "type": "object"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "client_request_id",
          "query",
          "title",
          "status",
          "created_at"
        ],
        "title": "TaskSnapshot",
        "type": "object"
      },
...
      "UpdateDeepSeekSettingsRequest": {
        "properties": {
          "api_key": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Api Key",
            "type": "string"
          }
        },
        "required": [
          "api_key"
        ],
        "title": "UpdateDeepSeekSettingsRequest",
        "type": "object"
      },
      "UpdateTaskRequest": {
        "properties": {
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "UpdateTaskRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "多Agent学术研究助手",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/research/tasks": {
      "post": {
        "operationId": "create_task_api_v1_research_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResearchTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Task",
        "tags": [
          "research tasks"
        ]
      }
    },
    "/api/v1/research/tasks/active": {
      "get": {
        "operationId": "get_active_task_api_v1_research_tasks_active_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "204": {
            "description": "No active task"
          }
        },
        "summary": "Get Active Task",
        "tags": [
          "research tasks"
        ]
      }
    },
    "/api/v1/research/tasks/{task_id}": {
      "delete": {
        "operationId": "delete_task_api_v1_research_tasks__task_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Task",
        "tags": [
          "research tasks"
        ]
      },
      "get": {
        "operationId": "get_task_api_v1_research_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task",
        "tags": [
          "research tasks"
        ]
      },
      "patch": {
        "operationId": "update_task_api_v1_research_tasks__task_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Task",
        "tags": [
          "research tasks"
        ]
      }
    },
    "/api/v1/research/tasks/{task_id}/cancel": {
      "post": {
        "operationId": "cancel_task_api_v1_research_tasks__task_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Task",
        "tags": [
          "research tasks"
        ]
      }
    },
    "/api/v1/research/tasks/{task_id}/events": {
      "get": {
        "operationId": "stream_events_api_v1_research_tasks__task_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Events",
        "tags": [
          "research events"
        ],
        "x-event-schema": {
          "$ref": "#/components/schemas/ResearchEvent"
        }
      }
    },
...
    "/api/v1/research/tasks/{task_id}/result": {
      "get": {
        "operationId": "get_result_api_v1_research_tasks__task_id__result_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchTaskResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Result",
        "tags": [
          "research results"
        ]
      }
    },
    "/api/v1/research/tasks/{task_id}/retry": {
      "post": {
        "operationId": "retry_task_api_v1_research_tasks__task_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Task",
        "tags": [
          "research tasks"
        ]
      }
    },
    "/api/v1/settings/deepseek": {
      "get": {
        "operationId": "get_deepseek_settings_api_v1_settings_deepseek_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeepSeekSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Deepseek Settings",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "update_deepseek_settings_api_v1_settings_deepseek_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDeepSeekSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeepSeekSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Deepseek Settings",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}